requirements.txt
transformers>=4.57.0
torch>=2.9.0
torchaudio>=2.9.0
accelerate>=1.12.0
librosa>=0.11.0
soundfile>=0.13.0
sounddevice>=0.4.6
numpy>=1.24.0
assemblyai>=0.17.0
python-dotenv>=1.0.0
pyaudio>=0.2.13
datasets>=2.14.0,<4.0  # Pin to <4.0 to avoid torchcodec dependency
torchmetrics>=1.8.0
mecab-python3>=1.0.9
unidic-lite>=1.0.8
